+++ gcp/provider.go
@@ -2,6 +2,7 @@ package gcp
 
 import (
 	"context"
+	"io/ioutil"
 	"log"
 	"regexp"
 	"strconv"
@@ -9,6 +10,7 @@ import (
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/hashicorp/terraform/helper/validation"
 	"github.com/hashicorp/terraform/terraform"
+	"golang.org/x/oauth2"
 	"golang.org/x/oauth2/google"
 	"google.golang.org/api/cloudresourcemanager/v1"
 )
@@ -37,6 +39,12 @@ func Provider() terraform.ResourceProvider {
 				DefaultFunc: schema.EnvDefaultFunc("GCP_CREDENTIALS", nil),
 				Description: "The credentials for GCP API operations.",
 			},
+			"token_duration": {
+				Type:        schema.TypeInt,
+				Optional:    true,
+				DefaultFunc: schema.EnvDefaultFunc("GCP_TOKEN_DURATION", nil),
+				Description: "The token duration in minutes from service account name for GCP API operations.",
+			},
 		},
 
 		ResourcesMap: map[string]*schema.Resource{
@@ -56,9 +64,42 @@ func Provider() terraform.ResourceProvider {
 	}
 }
 
-func getProjectNumber(p string) (string, error) {
+func getProjectNumber(p string, d *schema.ResourceData) (string, error) {
 	ctx := context.Background()
-	c, err := google.DefaultClient(ctx, cloudresourcemanager.CloudPlatformScope)
+	var ts *google.Credentials
+	var b []byte
+
+	re := regexp.MustCompile(`^[[a-z]([-a-z0-9]*[a-z0-9])@[a-z0-9-]+\.iam\.gserviceaccount\.com$`)
+
+	if v, ok := d.GetOk("service_account"); ok {
+		serviceAccount := v.(string)
+		if re.MatchString(serviceAccount) {
+			cloudresourcemanagerService, err := cloudresourcemanager.NewService(ctx)
+			if err != nil {
+				log.Printf("getProjectNumber: Cannot get cloud resource manager service(%s)", err)
+				return "", err
+			}
+			resp, err := cloudresourcemanagerService.Projects.Get(p).Context(ctx).Do()
+			if err != nil {
+				log.Printf("getProjectNumber: Cannot find project number (%s)", err)
+				return "", err
+			}
+			return strconv.FormatInt(resp.ProjectNumber, 10), nil
+		}
+		var err error
+		b, err = ioutil.ReadFile(serviceAccount)
+		if err != nil {
+			return "", err
+		}
+	} else if v, ok := d.GetOk("credentials"); ok {
+		b = []byte(v.(string))
+	}
+
+	ts, err := google.CredentialsFromJSON(ctx, b, cloudresourcemanager.CloudPlatformScope)
+	if err != nil {
+		return "", err
+	}
+	c := oauth2.NewClient(ctx, ts.TokenSource)
 	if err != nil {
 		log.Printf("getProjectNumber: Not able to get client (%s)", err)
 		return "", err
@@ -81,6 +122,8 @@ func providerConfigure(d *schema.ResourceData) (interface{}, error) {
 	project := d.Get("project").(string)
 	serviceAccount := d.Get("service_account").(string)
 	credentials := d.Get("credentials").(string)
+	tokenDuration := d.Get("token_duration").(int)
+
 	// check if project is project number or project ID
 	// project number is a string with numbers
 	// project ID must be 6 to 30 lowercase letters, digits, or hyphens. It must start with a letter. Trailing hyphens are prohibited
@@ -88,11 +131,12 @@ func providerConfigure(d *schema.ResourceData) (interface{}, error) {
 	if !isProjectNumber {
 		isProjectID, err := regexp.MatchString("^[a-z][a-z0-9-]+[a-z0-9]+$", project)
 		if isProjectID {
-			projectNumber, err = getProjectNumber(project)
+			projectNumber, err = getProjectNumber(project, d)
 			if err != nil {
 				log.Printf("providerConfigure: Cannot find project number (%s)", err)
 				return nil, err
 			}
+			log.Printf("**** Project number %v", projectNumber)
 		} else {
 			log.Printf("providerConfigure: Project %s format is not correct. It should be either numerical project number or project ID in xxx-xxx-xxxx format.", project)
 			return nil, err
@@ -104,6 +148,7 @@ func providerConfigure(d *schema.ResourceData) (interface{}, error) {
 		Project:        projectNumber,
 		ServiceAccount: serviceAccount,
 		Credentials:    credentials,
+		TokenDuration:  tokenDuration,
 	}
 
 	return config.clientFun()