+++ gcp/resource_netapp_gcp_volume.go
@@ -320,6 +320,10 @@ func resourceGCPVolume() *schema.Resource {
 				Optional: true,
 				Default:  true,
 			},
+			"pool_id": {
+				Type:     schema.TypeString,
+				Optional: true,
+			},
 			"smb_share_settings": {
 				Type:     schema.TypeSet,
 				Optional: true,
@@ -403,6 +407,10 @@ func resourceGCPVolumeCreate(d *schema.ResourceData, meta interface{}) error {
 
 	if v, ok := d.GetOk("volume_path"); ok {
 		volume.CreationToken = v.(string)
+	}
+
+	if v, ok := d.GetOk("pool_id"); ok {
+		volume.PoolID = v.(string)
 	}
 
 	var volType string
@@ -649,6 +657,9 @@ func resourceGCPVolumeRead(d *schema.ResourceData, meta interface{}) error {
 	if err := d.Set("service_level", slevel); err != nil {
 		return fmt.Errorf("Error reading volume service_level: %s", err)
 	}
+	if err := d.Set("pool_id", res.PoolID); err != nil {
+		return fmt.Errorf("Error reading volume pool_id: %s", err)
+	}
 	for i, protocol := range res.ProtocolTypes {
 		if protocol == "CIFS" {
 			res.ProtocolTypes[i] = "SMB"
@@ -740,7 +751,7 @@ func resourceGCPVolumeRead(d *schema.ResourceData, meta interface{}) error {
 	}
 	if _, ok := d.GetOk("unix_permissions"); ok {
 		if err := d.Set("unix_permissions", res.UnixPermissions); err != nil {
-			return fmt.Errorf("Error reading volume unix_permissions: %S", err)
+			return fmt.Errorf("Error reading volume unix_permissions: %s", err)
 		}
 	}
 	if _, ok := d.GetOk("billing_label"); ok {